pyHR/requirements.txt
fastapi[standard]==0.115.12
sqlmodel>=0.0.24
psycopg2-binary>=2.9.10
loguru==0.7.3
alembic==1.15.2
SQLAlchemy~=2.0.40
python-dotenv~=1.1.0
pyjwt>=2.10.1
bcrypt==4.0.1
passlib[bcrypt]
python-multipart